tests migrated to pytest
