+++ tests/test_loader.py
@@ -1,22 +1,5 @@
 import pytest
-from emva1288.camera.dataset_generator import DatasetGenerator
-from emva1288.process.parser import ParseEmvaDescriptorFile
 from emva1288.process.loader import LoadImageData
-
-
-# @pytest.fixture
-# def dataset():
-#     dataset = DatasetGenerator(height=50,
-#                                width=100,
-#                                bit_depth=8,
-#                                L=50,
-#                                steps=10)
-#     descriptor_path = dataset.descriptor_path
-#     # create the parser
-#     parser = ParseEmvaDescriptorFile(descriptor_path)
-#     # create loader
-#     loader = LoadImageData(parser.images)
-#     return dataset, parser, loader
 
 
 @pytest.mark.parametrize("dataset", ['multi_exposure'], indirect=True)
